+++ src/Spd.Manager.Licence/BizLicenceAppMananger.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Spd.Manager.Licence;
+internal class BizLicenceAppMananger
+{
+}
