+++ src/Spd.Manager.Licence/FeeManager.cs
@@ -1,0 +1,37 @@
+﻿using AutoMapper;
+using MediatR;
+using Microsoft.Extensions.Logging;
+using Spd.Resource.Applicants.LicenceApplication;
+using Spd.Resource.Applicants.LicenceFee;
+
+namespace Spd.Manager.Licence;
+internal class FeeManager :
+        IRequestHandler<GetLicenceFeeListQuery, LicenceFeeListResponse>,
+        IFeeManager
+{
+    private readonly ILicenceFeeRepository _licenceFeeRepository;
+    private readonly ILogger<IFeeManager> _logger;
+    private readonly IMapper _mapper;
+
+    public FeeManager(
+        ILicenceFeeRepository licenceFeeRepository,
+        ILogger<IFeeManager> logger,
+        IMapper mapper)
+    {
+        _licenceFeeRepository = licenceFeeRepository;
+        _mapper = mapper;
+        _logger = logger;
+    }
+
+    public async Task<LicenceFeeListResponse> Handle(GetLicenceFeeListQuery query, CancellationToken ct)
+    {
+        var fees = await _licenceFeeRepository.GetLicenceFeeAsync((WorkerLicenceTypeEnum)query.WorkerLicenceTypeCode, ct);
+        var feeResps = _mapper.Map<IEnumerable<LicenceFeeResponse>>(fees.LicenceFees);
+
+        return new LicenceFeeListResponse
+        {
+            LicenceFees = feeResps
+
+        };
+    }
+}
