+++ src/Spd.Presentation.Licensing/Controllers/ConfigurationController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
-using Spd.Manager.Cases.Licence;
+using Spd.Manager.Licence;
 using Spd.Utilities.LogonUser.Configurations;
 using Spd.Utilities.Recaptcha;
 using Spd.Utilities.Shared;
