+++ src/Spd.Presentation.Licensing/Controllers/WorkerLicensingController.cs
@@ -2,7 +2,7 @@ using FluentValidation;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Spd.Manager.Cases.Licence;
+using Spd.Manager.Licence;
 using Spd.Manager.Membership.UserProfile;
 using Spd.Presentation.Licensing.Configurations;
 using Spd.Presentation.Licensing.Services;
