+++ src/Spd.Presentation.Licensing/Program.cs
@@ -7,7 +7,6 @@ using Spd.Utilities.Dynamics;
 using Spd.Utilities.FileStorage;
 using Spd.Utilities.Hosting;
 using Spd.Utilities.LogonUser;
-using Spd.Utilities.Payment;
 using Spd.Utilities.TempFileStorage;
 using System.Reflection;
 using System.Security.Principal;
@@ -55,7 +54,7 @@ builder.Services.AddTransient<IMultipartRequestService, MultipartRequestService>
 builder.Services.AddFileStorageProxy(builder.Configuration);
 builder.Services
   .AddBCeIDService(builder.Configuration)
-  .AddPaymentService(builder.Configuration)
+  // .AddPaymentService(builder.Configuration)
   .AddDynamicsProxy(builder.Configuration)
   .AddAddressAutoComplete(builder.Configuration);
 builder.Services.ConfigureComponentServices(builder.Configuration, builder.Environment, assemblies);
