+++ src/Spd.Presentation.Licensing/Spd.Presentation.Licensing.csproj
@@ -29,7 +29,7 @@
 
 	<ItemGroup>
 	  <ProjectReference Include="..\Spd.Manager.Admin\Spd.Manager.Admin.csproj" />
-	  <ProjectReference Include="..\Spd.Manager.Cases\Spd.Manager.Cases.csproj" />
+	  <ProjectReference Include="..\Spd.Manager.Licence\Spd.Manager.Licence.csproj" />
 	  <ProjectReference Include="..\Spd.Manager.Membership\Spd.Manager.Membership.csproj" />
 	  <ProjectReference Include="..\Spd.Utilities.Hosting\Spd.Utilities.Hosting.csproj" />
 	  <ProjectReference Include="..\Spd.Utilities.LogonUser\Spd.Utilities.LogonUser.csproj" />
