+++ src/Spd.sln
@@ -65,6 +65,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Spd.Utilities.Payment", "Sp
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Spd.Utilities.BCeIDWS", "Spd.Utilities.BCeIDWS\Spd.Utilities.BCeIDWS.csproj", "{C16D353E-67D6-4EA0-9400-167234A5DC35}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Spd.Manager.Licence", "Spd.Manager.Licence\Spd.Manager.Licence.csproj", "{119B3695-6918-46DD-82B7-493C9F6CAB77}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -163,6 +165,10 @@ Global
 		{C16D353E-67D6-4EA0-9400-167234A5DC35}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C16D353E-67D6-4EA0-9400-167234A5DC35}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C16D353E-67D6-4EA0-9400-167234A5DC35}.Release|Any CPU.Build.0 = Release|Any CPU
+		{119B3695-6918-46DD-82B7-493C9F6CAB77}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{119B3695-6918-46DD-82B7-493C9F6CAB77}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{119B3695-6918-46DD-82B7-493C9F6CAB77}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{119B3695-6918-46DD-82B7-493C9F6CAB77}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -191,6 +197,7 @@ Global
 		{C71B163E-216F-481E-8402-B6BEBF7EFDBC} = {744553C0-C41C-4315-9592-92A97938234E}
 		{F0BDD571-19CF-41F8-91B6-2115109DDC2B} = {744553C0-C41C-4315-9592-92A97938234E}
 		{C16D353E-67D6-4EA0-9400-167234A5DC35} = {744553C0-C41C-4315-9592-92A97938234E}
+		{119B3695-6918-46DD-82B7-493C9F6CAB77} = {AFA7034A-A8B1-489C-9A24-030B05C5412B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {07609A1C-A62A-4B0C-85C1-40600CB3DC45}
